requirements.txt
aiofiles==24.1.0
aiogram==3.15.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
aiosqlite==0.22.1
alembic==1.18.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
bcrypt==4.2.1
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
cryptography==43.0.3
fastapi==0.115.5
frozenlist==1.5.0
greenlet==3.3.0
h11==0.14.0
idna==3.10
Jinja2==3.1.4
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.1.0
passlib==1.7.4
propcache==0.2.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.10.0
python-dotenv==1.0.1
python-multipart==0.0.17
redis==7.1.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1
yarl==1.18.0
